backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
pyotp==2.9.0
apscheduler==3.10.4
requests==2.31.0
aiofiles==23.2.1
pillow==10.1.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
python-dotenv==1.0.0
matplotlib==3.8.2
plotly==5.17.0
pandas==2.1.4
numpy==1.26.2
opencv-python-headless==4.8.1.78
pyzbar==0.1.9
email-validator==2.1.0